+++ src/handlers/command.py
@@ -19,10 +19,12 @@ def __watch(bot, update):
     splitedMsg = text.split()
     command = cut_command(splitedMsg[0][1:])
     args = splitedMsg[1:]
+
     # logging.info("UPDATE: " + str(update))
     # logging.info("idUser: " + str(usr_id))
     # logging.info("Command: " + command)
     # logging.info("Arguments: " + ", ".join(args))
+
     sender = message.chat.get_member(usr_id)
     logging.info(sender)
     logging.info(sender.status)
@@ -35,6 +37,7 @@ def __watch(bot, update):
             get_banned_words(update, args)
         elif command == "free":
             free_word(update, args)
+
     else:
         user_unauthorized(update)
